refactor(wizard): #1318 move the no-JavaScript form fallback into its own module
`wizard.py` sits at 674 against a 674 ceiling in `docs/dev/file-budget.tsv`,
and ceilings only ever go down — re-derived at `scripts/lint-file-budget.sh`,
where exactly one row is exempt and it is not this one. There is also no width
to reclaim: ruff's line-length is 100 and the file already has 99 lines in the
90-99 column band, so reflowing comments buys nothing. #1318's wizard half
needs roughly eleven lines there (a saved-role parse, one key in the
`/api/wizard-state` payload, a small `keep-role` POST handler and its route),
so the space has to come from an extraction. This is that extraction, kept as
its own commit so it can be read as the no-op it is.

`build_config` was the largest self-contained thing in the file: a pure mapping
from form fields to a pithead config, with no module state, no spool access and
no request handling. `wizard.py` re-imports the name, so `wizard.build_config`
still resolves for the one call site in `submit` and for the eleven assertions
in `dashboard/tests/web/test_wizard.py` — which is why that file, itself at
974/974 with no headroom, needs no edit at all.

`wizard.py` 674 -> 608. The ceiling is deliberately left at 674 rather than
regenerated down to 608: lowering a row reds every un-rebased in-flight branch,
and the headroom is what the next commit spends.

RUN:
  - The moved 66 lines are BYTE-IDENTICAL to lines 256-321 of `wizard.py` at
    the parent commit, compared with `cmp` — this is a move, not a rewrite.
  - `wizard.build_config` still resolves and still behaves: called through the
    `wizard` module it returns the wallet, the non-default timezone under
    `dashboard`, and the paired telegram block, and reports its new home as
    `mining_dashboard.wizard_form`.
  - `make lint-py` rc 0 at the gate's own pinned ruff, both `check` and
    `format --check` (224 files). My first run of it reported rc 0 from a
    pipeline whose last element was `tail`; the rc above is the make rc.
  - `make lint-file-budget` rc 0. `lane-guard` rc 0.

NOT RUN: the dashboard pytest suite — the appliance lane holds this box for a
KVM battery pre-flight until ~00:25Z. CI covers it, and `test_wizard.py` is
untouched by this commit.

COORDINATED: `lane-dashboard-46` confirmed in its own pane that it has touched
neither `wizard.py` nor `wizard.mjs`, and that #1855/#1848 — the Tari/XvB
question block, which would land in `build_config` — has not started, so this
extraction lands first and its successor rebases onto it rather than around it.
